firmware/firmware_MIN/bridge_usb/requirements.txt
# AGUADA Bridge USB - Dependências Python

pyserial>=3.5
requests>=2.28.0
